+++ console/src/components/Services/Data/Add/TableComment.js
@@ -2,16 +2,15 @@ import React from 'react';
 
 import PropTypes from 'prop-types';
 
-import TextInput from '../../../Common/TextInput/TextInput';
-
 const TableComment = ({ onChange }) => {
   return [
-    <TextInput
+    <textarea
       key="add_table_comment_element"
       type="text"
       placeholder="comment"
       data-test="tableComment"
       onChange={onChange}
+      className="form-control"
     />,
   ];
 };