+++ console/src/components/Services/Data/TableModify/TableCommentEditor.js
@@ -41,9 +41,12 @@ const TableCommentEditor = ({
   if (tableComment && !tableCommentEdit.enabled) {
     commentHtml = (
       <div>
-        <div className="rounded bg-secondary-light border border-gray-300 border-l-4 border-l-secondary py-sm px-md mb-sm">
+        <textarea
+          className="w-100 rounded bg-secondary-light border border-gray-300 border-l-4 border-l-secondary py-sm px-md mb-sm"
+          readOnly
+        >
           {tableComment}
-        </div>
+        </textarea>
         {!readOnly && (
           <div
             onClick={editCommentClicked}
@@ -58,13 +61,14 @@ const TableCommentEditor = ({
   } else if (tableCommentEdit.enabled) {
     commentHtml = (
       <div className="flex items-center">
-        <input
+        <textarea
           onChange={commentEdited}
           className="form-control"
           type="text"
-          value={tableCommentEdit.value}
           defaultValue={tableComment}
-        />
+        >
+          {tableCommentEdit.value}
+        </textarea>
         <div onClick={commentEditSave} className="ml-sm cursor-pointer">
           Save
         </div>